\name{data}
\alias{data}
\alias{rnaseq}
\non_function{}


\title{
	example of RNA-Seq data
}
\description{
	RNA-Seq data generated from next-generation sequencing technology.
}
\usage{
	data(rsgcc)
}



\details{
	This RNA-Seq data was obtained from [1].
}


\references{
      [1] Davidson R.M., Hansey, C. N., Gowda, M. et al. Utility of RNA sequencing for analysis of maize reproductive transcriptiomes. The Plant Genome, 2011, 4(3): 191-203.
}


\examples{
   data(rsgcc)
   x <- rnaseq[1:10,] #Just use a small subset of RNA-Seq data 
}


\keyword{
 datasets
}
