\name{data}
\alias{data}
\alias{rnaseq}
\non_function{}


\title{
	example of RNA-Seq gene expression data 
}
\description{
	RNA-Seq profiled gene expression data of 100 genes
}
\usage{
	data(rsgcc)
}


\examples{
   data(rsgcc)
   x <- rnaseq[1:3,] #The first 3 genes in GEMatrix "rnaseq". 
}


\keyword{
 datasets
}
